+++ Lab-06-solution/src/main/java/com/petclinic/core/Visit.java
@@ -10,7 +10,7 @@ import java.time.LocalDate;
 @Entity
 public class Visit {
     @Id @GeneratedValue
-    private int id;
+    private Long id;
     @Column(unique = true)
     private String referenceNumber;
     private LocalDate date;
@@ -21,14 +21,14 @@ public class Visit {
     public Visit() {
     }
 
-    public Visit(int id, String referenceNumber, LocalDate date, String purpose) {
+    public Visit(Long id, String referenceNumber, LocalDate date, String purpose) {
         this.id = id;
         this.referenceNumber = referenceNumber;
         this.date = date;
         this.purpose = purpose;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
